pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faster_app"
version = "0.1.3"
description = "一个轻量级的 Python Web 框架, 提供自动发现、模型基类、命令行工具等功能"
readme = "README.md"
license = "MIT"
authors = [{ name = "peizhenfei", email = "peizhenfei@hotmail.com" }]
maintainers = [{ name = "peizhenfei", email = "peizhenfei@hotmail.com" }]
keywords = ["web", "framework", "fastapi", "tortoise", "orm", "cli"]

requires-python = ">=3.12"
dependencies = [
  "aerich>=0.9.1",
  "asyncpg>=0.30.0",
  "fastapi>=0.116.1",
  "fastapi-pagination>=0.14.1",
  "fire>=0.7.1",
  "pydantic>=2.11.7",
  "pydantic-settings>=2.10.1",
  "rich>=14.1.0",
  "tortoise-orm[asyncpg]>=0.25.1",
  "uvicorn>=0.35.0",
]


[project.urls]
Homepage = "https://github.com/mautops/faster_app"
Documentation = "https://github.com/mautops/faster_app#readme"
Repository = "https://github.com/mautops/faster_app.git"
"Bug Tracker" = "https://github.com/mautops/faster_app/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["faster_app*"]
exclude = ["tests*", "docs*", "__pycache__*", "*.pyc", "*.pyo"]

[tool.setuptools.package-data]
"faster_app" = ["statics/*", ".env.example", "runtime/*"]

[project.scripts]
faster = "faster_app.cli:main"

[dependency-groups]
dev = [
  "build>=1.3.0",
  "mkdocs>=1.6.1",
  "mkdocs-material>=9.6.21",
  "mkdocstrings[python]>=0.30.1",
  "twine>=6.2.0",
  "ruff>=0.8.0",
  "mypy>=1.14.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
]

[tool.ruff]
# Ruff configuration for code quality
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# Enable recommended rules plus additional checks
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "N",   # pep8-naming
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "SIM", # flake8-simplify
]
ignore = [
  "E501", # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # Allow unused imports in __init__.py

[tool.mypy]
# MyPy configuration for type checking
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
# Pytest configuration
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--strict-markers",
  "--cov=faster_app",
  "--cov-report=term-missing",
  "--cov-report=html",
]
asyncio_mode = "auto"
